Add AuthController and move login/signup features
