+++ backend/app.js
@@ -1,6 +1,7 @@
 //IMPORT PACKAGE
 const express = require("express");
 const userRouter = require("./routes/user");
+const authRouter = require("./routes/auth");
 const otpRouter = require("./routes/otp");
 const cors = require("cors");
 
@@ -16,5 +17,6 @@ app.use(cors());
 //USING ROUTES
 app.use("/api/user", userRouter);
 app.use("/api/otp", otpRouter);
+app.use("/api/auth", authRouter);
 
 module.exports = app;
