+++ backend/routes/auth.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const authController = require("../controllers/authController");
+
+const router = express.Router();
+
+router.post("/login", authController.login);
+router.post("/signup", authController.signup);
+
+module.exports = router;
